requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
deepspeed>=0.9.0
diffusers>=0.21.0
ray[train]>=2.6.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0
safetensors>=0.3.0
lmdb>=1.4.0
tqdm>=4.65.0
psutil>=5.9.0
numpy>=1.24.0
scipy>=1.10.0
biopython>=1.81
peft>=0.4.0
einops>=0.6.0 